+++ src/api/option-min.ts
@@ -1,5 +1,6 @@
 import { MinOption } from "./bundlib-options";
 import { MinifyOptions, ModuleOutputFields } from "./pkg-analized";
+import setProp from "./set-prop";
 import { isArray, isBool, isNull } from "./type-check";
 
 export function isOutputField(value: any): value is ModuleOutputFields {
@@ -17,23 +18,14 @@ export function isValidMinOption(value: any): value is MinOption {
 }
 
 export function baseMinOption(value?: boolean): MinifyOptions {
-  return (["main", "module", "browser"] as ModuleOutputFields[]).reduce((options, field) => {
-    options[field] = !!value;
-    return options;
-  }, {} as Record<ModuleOutputFields, boolean>);
+  return (["main", "module", "browser"] as ModuleOutputFields[]).reduce((options, field) => (
+    setProp(field, !!value, options)
+  ), {} as Record<ModuleOutputFields, boolean>);
 }
 
 export function normalizeMin(min: MinOption): MinifyOptions {
-  return !min
-    ? baseMinOption()
-    : min === true
-      ? baseMinOption(true)
-      : isArray(min)
-        ? min.reduce((result, value) => {
-          result[value] = true;
-          return result;
-        }, baseMinOption())
-        : Object.assign<MinifyOptions, Partial<MinifyOptions>>(baseMinOption(), {
-          [min]: true,
-        });
+  return !min ? baseMinOption()
+    : min === true ? baseMinOption(true)
+      : isArray(min) ? min.reduce((result, field) => setProp(field, true, result), baseMinOption())
+        : setProp(min, true, baseMinOption());
 }